Extracted ICoordinateConverter
