+++ OutlinesApp/OutlinesApp.csproj
@@ -1,4 +1,4 @@
-<?xml version="1.0" encoding="utf-8"?>
+﻿<?xml version="1.0" encoding="utf-8"?>
 <Project ToolsVersion="15.0" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
   <Import Project="$(MSBuildExtensionsPath)\$(MSBuildToolsVersion)\Microsoft.Common.props" Condition="Exists('$(MSBuildExtensionsPath)\$(MSBuildToolsVersion)\Microsoft.Common.props')" />
   <PropertyGroup>
@@ -90,7 +90,9 @@
     <Compile Include="Services\CachedCoordinateConverter.cs" />
     <Compile Include="Services\IColorPickerService.cs" />
     <Compile Include="Services\InputMaskingService.cs" />
+    <Compile Include="Services\ICoordinateConverter.cs" />
     <Compile Include="Services\IScreenHelper.cs" />
+    <Compile Include="Services\ScreenHelper.cs" />
     <Compile Include="Services\ServiceContainer.cs" />
     <Compile Include="SnapshotInspectorWindow.xaml.cs">
       <DependentUpon>SnapshotInspectorWindow.xaml</DependentUpon>
@@ -120,7 +122,7 @@
     <Compile Include="Views\OutlinesOverlay.xaml.cs">
       <DependentUpon>OutlinesOverlay.xaml</DependentUpon>
     </Compile>
-    <Compile Include="Services\ScreenHelper.cs" />
+    <Compile Include="Services\LiveCoordinateConverter.cs" />
     <Compile Include="Views\PropertiesPanel.xaml.cs">
       <DependentUpon>PropertiesPanel.xaml</DependentUpon>
     </Compile>
