+++ OutlinesApp/Services/ICoordinateConverter.cs
@@ -1,0 +1,14 @@
+﻿using System.Windows;
+
+namespace OutlinesApp.Services
+{
+    public interface ICoordinateConverter
+    {
+        Point PointFromScreen(Point screenPoint);
+        Point PointToScreen(Point localPoint);
+        Rect RectFromScreen(Rect screenRect);
+        Rect RectToScreen(Rect localRect);
+        Size SizeFromScreen(Size screenSize);
+        Size SizeToScreen(Size localSize);
+    }
+}
