+++ OutlinesApp/Services/IScreenHelper.cs
@@ -5,11 +5,5 @@ namespace OutlinesApp.Services
     public interface IScreenHelper
     {
         Rect GetMonitorRect(Point point);
-        Point PointFromScreen(Point screenPoint);
-        Point PointToScreen(Point localPoint);
-        Rect RectFromScreen(Rect screenRect);
-        Rect RectToScreen(Rect localRect);
-        Size SizeFromScreen(Size screenSize);
-        Size SizeToScreen(Size localSize);
     }
 }
